Add things to the .gitignore file
I generated it with the command `dotnet new gitignore` and added a
couple of ignores back that were present before.

The reason to this change was because I opened the project in Visual
Studio on Windows, which created a folder `.vs/` which wasn't previously
ignored. So this change fixes it, and ignores other potential files that
shouldn't be committed.